pyproject.toml
[project]
name = "test-architecture-learning"
version = "0.1.0"
description = "Roadmap de aprendizaje de 12 meses para arquitectura de pruebas"
authors = [
  {name = "Ignacio PS", email = "contacto@ignaciops.dev"}
]
requires-python = ">=3.14"
dependencies = [
    "playwright>=1.40",
    "pytest>=9.0",
    "pytest-playwright>=0.7.1",
    "allure-pytest>=2.15.0",
]

[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["framework*", "tests*"]
exclude = ["months*", "suts*", "docs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
  "smoke: Smoke tests",
  "e2e: End-to-end tests",
  "api: API tests",
  "critical: Critical path tests"
]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--alluredir=allure-results",
    # "--screenshot=only-on-failure",
    # "--headed", # Quitar comentario para ver navegador durante ejecución de pruebas
    # "--slowmo=100", # Quitar comentario para ralentizar la ejecución de pruebas
]